refactor(frontend/db): 🚚 move db to /lib/
