--- app/frontend/src/features/db/data.ts
@@ -1,21 +1,0 @@
-export const users = [
-	{
-		id: "1",
-		name: "Biran",
-		message: "Ich liebe sie! Hello",
-		timestamp: "21:15",
-	},
-	{
-		id: "2",
-		name: "Isolde",
-		message: "Klingt nach einer Kunstform.",
-		timestamp: "22:44",
-	},
-	{
-		id: "3",
-		name: "Gustav",
-		message: "Ich liebe sie!",
-		timestamp: "21:15",
-		unreadCount: 2,
-	},
-];
